reorganized Autoscan files, seperated keyboard Controls from Autoscan, implemented basic functionality for arduino controller
