+++ Child Talker/App.config
@@ -2,6 +2,7 @@
 <configuration>
     <configSections>
         <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
+            <section name="Child_Talker.Properties.HardareIntegration" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
             <section name="Child_Talker.Properties.AutoscanSettings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
             <section name="Child_Talker.Properties.Autoscan" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
             <section name="Child_Talker.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
@@ -11,6 +12,11 @@
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.6.1" />
     </startup>
     <userSettings>
+        <Child_Talker.Properties.HardareIntegration>
+            <setting name="Active_COM_Port" serializeAs="String">
+                <value />
+            </setting>
+        </Child_Talker.Properties.HardareIntegration>
         <Child_Talker.Properties.AutoscanSettings>
             <setting name="Enabled" serializeAs="String">
                 <value>False</value>
