+++ Child Talker/Child Talker.csproj
@@ -98,6 +98,16 @@
     <Reference Include="WindowsFormsIntegration" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Properties\HardareIntegration.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTimeSharedInput>True</DesignTimeSharedInput>
+      <DependentUpon>HardareIntegration.settings</DependentUpon>
+    </Compile>
+    <Compile Include="Utilities\HardwareIntegrations\ArduinoIntegration.cs" />
+    <Compile Include="Utilities\Autoscan\DefaultBehavior.cs" />
+    <Compile Include="Utilities\Autoscan\DependencyProperties.cs" />
+    <Compile Include="Utilities\Autoscan\EventHandling.cs" />
+    <Compile Include="Utilities\HardwareIntegrations\KeyboardIntegration.cs" />
     <Compile Include="Utilities\MessageBox.cs" />
     <Compile Include="Utilities\MessageBoxWindow.xaml.cs">
       <DependentUpon>MessageBoxWindow.xaml</DependentUpon>
@@ -136,7 +146,7 @@
     <Compile Include="TalkerViews\SettingPage\SettingsPage.xaml.cs">
       <DependentUpon>SettingsPage.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Utilities\Autoscan2.cs" />
+    <Compile Include="Utilities\Autoscan\Autoscan2.cs" />
     <Compile Include="Utilities\ConsoleControls.cs" />
     <Compile Include="TalkerViews\EnvControlsPage\EnvControls.xaml.cs">
       <DependentUpon>EnvControls.xaml</DependentUpon>
@@ -284,6 +294,10 @@
     <None Include="Properties\AutoscanSettings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>AutoscanSettings.Designer.cs</LastGenOutput>
+    </None>
+    <None Include="Properties\HardareIntegration.settings">
+      <Generator>SettingsSingleFileGenerator</Generator>
+      <LastGenOutput>HardareIntegration.Designer.cs</LastGenOutput>
     </None>
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
