+++ Child Talker/Properties/HardareIntegration.settings
@@ -1,0 +1,9 @@
+﻿<?xml version='1.0' encoding='utf-8'?>
+<SettingsFile xmlns="http://schemas.microsoft.com/VisualStudio/2004/01/settings" CurrentProfile="(Default)" GeneratedClassNamespace="Child_Talker.Properties" GeneratedClassName="HardareIntegration">
+  <Profiles />
+  <Settings>
+    <Setting Name="Active_COM_Port" Type="System.String" Scope="User">
+      <Value Profile="(Default)" />
+    </Setting>
+  </Settings>
+</SettingsFile>
