+++ Child Talker/Resources/ColorSchemes/BaseScheme.xaml
@@ -3,7 +3,8 @@
                     xmlns:custom="clr-namespace:Child_Talker"
                     xmlns:utilities="clr-namespace:Child_Talker.Utilities"
                     xmlns:tb="clr-namespace:Child_Talker.TalkerButton"
-                    xmlns:properties="clr-namespace:Child_Talker.Properties">
+                    xmlns:properties="clr-namespace:Child_Talker.Properties"
+                    xmlns:autoscan="clr-namespace:Child_Talker.Utilities.Autoscan">
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="SystemColors.xaml"></ResourceDictionary>
     </ResourceDictionary.MergedDictionaries>
@@ -32,12 +33,12 @@
     <Style TargetType="{x:Type TextBox}" BasedOn="{StaticResource genTextBox}"/>
 
     <Style TargetType="Panel" x:Key="PanelTriggers">
-        <Setter Property="utilities:Autoscan2.IsHighlight" Value="False"></Setter>
+        <Setter Property="autoscan:Autoscan2.IsHighlight" Value="False"></Setter>
         <Style.Triggers>
-            <Trigger Property="utilities:Autoscan2.IsHighlight" Value="true">
+            <Trigger Property="autoscan:Autoscan2.IsHighlight" Value="true">
                 <Setter Property="Background" Value="{StaticResource ColorScheme.Background.Highlighted}"></Setter>
             </Trigger>
-            <Trigger Property="utilities:Autoscan2.IsHighlight" Value="False">
+            <Trigger Property="autoscan:Autoscan2.IsHighlight" Value="False">
                 <Setter Property="Background" Value="Transparent"></Setter>
             </Trigger>
         </Style.Triggers>
