+++ Child Talker/Resources/ColorSchemes/ScrollBarStyle.xaml
@@ -1,6 +1,7 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:util="clr-namespace:Child_Talker.Utilities">
+                    xmlns:util="clr-namespace:Child_Talker.Utilities"
+                    xmlns:autoscan="clr-namespace:Child_Talker.Utilities.Autoscan">
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="BaseScheme.xaml"></ResourceDictionary>
     </ResourceDictionary.MergedDictionaries>
@@ -300,16 +301,16 @@
             <ScrollBar x:Name="PART_HorizontalScrollBar" AutomationProperties.AutomationId="HorizontalScrollBar" Cursor="Arrow" Grid.Column="0" Maximum="{TemplateBinding ScrollableWidth}" Minimum="0" Orientation="Horizontal" Grid.Row="1" Visibility="{TemplateBinding ComputedHorizontalScrollBarVisibility}" Value="{Binding HorizontalOffset, Mode=OneWay, RelativeSource={RelativeSource TemplatedParent}}" ViewportSize="{TemplateBinding ViewportWidth}"/>
         </Grid>
         <ControlTemplate.Triggers>
-            <Trigger Property="util:Autoscan2.IsHighlight" Value="False">
+            <Trigger Property="autoscan:Autoscan2.IsHighlight" Value="False">
                 <Setter Property="Background" Value="Transparent"></Setter>
             </Trigger>
-            <Trigger Property="util:Autoscan2.IsHighlight" Value="True">
+            <Trigger Property="autoscan:Autoscan2.IsHighlight" Value="True">
                 <Setter Property="Background" Value="{StaticResource ColorScheme.Background.Highlighted}"></Setter>
             </Trigger>
         </ControlTemplate.Triggers>
     </ControlTemplate>
     <Style x:Key="genScrollViewer"  TargetType="{x:Type ScrollViewer}" >
-        <Setter Property="util:Autoscan2.IsHighlight" Value="False"></Setter>
+        <Setter Property="autoscan:Autoscan2.IsHighlight" Value="False"></Setter>
         <Setter Property="Template" Value="{StaticResource ScrollViewerControlTemplate}"/>
     </Style>
