+++ Child Talker/TalkerViews/SettingPage/SettingsTemplates.xaml
@@ -1,7 +1,8 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:local="clr-namespace:Child_Talker.TalkerViews.SettingPage"
-                    xmlns:utilities="clr-namespace:Child_Talker.Utilities">
+                    xmlns:utilities="clr-namespace:Child_Talker.Utilities"
+                    xmlns:autoscan="clr-namespace:Child_Talker.Utilities.Autoscan">
     <ResourceDictionary.MergedDictionaries>
         <ResourceDictionary Source="../../Resources/ColorSchemes/ButtonTemplates.xaml"></ResourceDictionary>
     </ResourceDictionary.MergedDictionaries>
@@ -46,7 +47,7 @@
                 <Setter TargetName="ButtonOuterBorder" Property="Margin" Value="20"></Setter>
                 <Setter TargetName="Ellipse" Property="Fill" Value="Transparent"></Setter>
             </Trigger>
-            <Trigger Property="utilities:Autoscan2.IsHighlight" Value="False">
+            <Trigger Property="autoscan:Autoscan2.IsHighlight" Value="False">
                 <Setter TargetName="ButtonOuterBorder" Property="Margin" Value="20"></Setter>
                 <Setter TargetName="Ellipse" Property="Fill" Value="Transparent"></Setter>
             </Trigger>
@@ -54,7 +55,7 @@
                 <Setter TargetName="ButtonOuterBorder" Property="Margin" Value="40"></Setter>
                 <Setter TargetName="Ellipse" Property="Fill" Value="{StaticResource ColorScheme.Background.Highlighted}"></Setter>
             </Trigger>
-            <Trigger Property="utilities:Autoscan2.IsHighlight" Value="True">
+            <Trigger Property="autoscan:Autoscan2.IsHighlight" Value="True">
                 <Setter TargetName="ButtonOuterBorder" Property="Margin" Value="40"></Setter>
                 <Setter TargetName="Ellipse" Property="Fill" Value="{StaticResource ColorScheme.Background.MouseHover}"></Setter>
             </Trigger>
